+++ src/main/java/insane96mcp/iguanatweaksreborn/module/experience/enchantments/enchantment/protection/ITRProtectionEnchantment.java
@@ -1,10 +1,6 @@
 package insane96mcp.iguanatweaksreborn.module.experience.enchantments.enchantment.protection;
 
 import insane96mcp.iguanatweaksreborn.module.experience.enchantments.EnchantmentsFeature;
-import insane96mcp.insanelib.InsaneLib;
-import insane96mcp.insanelib.world.enchantments.IEnchantmentTooltip;
-import net.minecraft.ChatFormatting;
-import net.minecraft.network.chat.Component;
 import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.item.ItemStack;
@@ -13,7 +9,7 @@ import net.minecraft.world.item.enchantment.EnchantmentCategory;
 import net.minecraft.world.item.enchantment.ProtectionEnchantment;
 import org.jetbrains.annotations.NotNull;
 
-public abstract class ITRProtectionEnchantment extends Enchantment implements IProtectionEnchantment, IEnchantmentTooltip {
+public abstract class ITRProtectionEnchantment extends Enchantment implements IProtectionEnchantment {
     public static final EquipmentSlot[] ARMOR_SLOTS = new EquipmentSlot[]{EquipmentSlot.HEAD, EquipmentSlot.CHEST, EquipmentSlot.LEGS, EquipmentSlot.FEET};
 
     public ITRProtectionEnchantment(Rarity rarity) {
@@ -55,9 +51,4 @@ public abstract class ITRProtectionEnchantment extends Enchantment implements IP
 
     @Override
     public abstract boolean isSourceReduced(DamageSource source);
-
-    @Override
-    public Component getTooltip(ItemStack stack, int lvl) {
-        return Component.translatable(this.getDescriptionId() + ".tooltip", InsaneLib.ONE_DECIMAL_FORMATTER.format(this.getDamageReduction(lvl) * 100f)).withStyle(ChatFormatting.DARK_PURPLE);
-    }
 }