+++ src/main/java/insane96mcp/iguanatweaksreborn/module/movement/stamina/Vigour.java
@@ -1,9 +1,6 @@
 package insane96mcp.iguanatweaksreborn.module.movement.stamina;
 
 import insane96mcp.insanelib.base.Feature;
-import insane96mcp.insanelib.world.enchantments.IEnchantmentTooltip;
-import net.minecraft.ChatFormatting;
-import net.minecraft.network.chat.Component;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.enchantment.Enchantment;
@@ -11,7 +8,7 @@ import net.minecraft.world.item.enchantment.EnchantmentCategory;
 
 import java.util.UUID;
 
-public class Vigour extends Enchantment implements IEnchantmentTooltip {
+public class Vigour extends Enchantment {
 
     public static final UUID MODIFIER_UUID = UUID.fromString("74e97c20-6a62-482f-b909-e709087b066a");
 
@@ -37,10 +34,5 @@ public class Vigour extends Enchantment implements IEnchantmentTooltip {
     @Override
     public boolean canApplyAtEnchantingTable(ItemStack stack) {
         return super.canApplyAtEnchantingTable(stack) && Feature.isEnabled(Stamina.class);
-    }
-
-    @Override
-    public Component getTooltip(ItemStack itemStack, int i) {
-        return Component.translatable(this.getDescriptionId() + ".tooltip").withStyle(ChatFormatting.DARK_PURPLE);
     }
 }